day 6

+++ ProjectMVC1/Models/CrsResult.cs
@@ -3,7 +3,7 @@
     public class CrsResult
     {
         public int CrsResultID { get; set; }
-        public int Degree { get; set; }
+        public double Degree { get; set; }
         public int CourseId { get; set; }
         public int TraineeId { get; set; }
         public Trainee Trainee { get; set; }
